+++ ldlidar_component/driver/include/tofbf.hpp
@@ -30,8 +30,8 @@ private:
   const int CONFIDENCE_LOW = 15;  // Low confidence threshold
   const int CONFIDENCE_SINGLE = 220;  // Discrete points require higher confidence
   const int SCAN_FRE = 4500;  // Default scan frequency, to change, read according to radar protocol
-  double offset_x, offset_y;
-  double curr_speed;
+  double _offset_x, _offset_y;
+  double _curr_speed;
   Tofbf() = delete;
   Tofbf(const Tofbf &) = delete;
   Tofbf & operator=(const Tofbf &) = delete;